episode tmdbToFluxArtwork

+++ app/src/main/java/com/kaem/flux/model/tmdb/TMDBEpisode.kt
@@ -10,7 +10,7 @@ data class TMDBEpisode(
     @SerializedName("runtime")
     val duration: Int,
     @SerializedName("episode_number")
-    val number: Float,
+    val number: Int,
     @SerializedName("season_number")
     val season: Int,
     @SerializedName("still_path")
